+++ actix-web-lab/src/lazy_data_shared.rs
@@ -1,28 +1,32 @@
 use std::{
-    cell::Cell,
     fmt,
     future::{Future, Ready, ready},
-    rc::Rc,
+    sync::Arc,
 };
 
 use actix_web::{Error, FromRequest, HttpRequest, dev, error};
-use futures_core::future::LocalBoxFuture;
-use tokio::sync::OnceCell;
+use futures_core::future::BoxFuture;
+use tokio::sync::{Mutex, OnceCell};
 use tracing::debug;
 
-/// A lazy extractor for thread-local data.
+/// A lazy extractor for globally shared data.
 ///
-/// Using `LazyData` as an extractor will not initialize the data; [`get`](Self::get) must be used.
-pub struct LazyData<T> {
-    inner: Rc<LazyDataInner<T>>,
+/// Unlike, [`LazyData`], this type implements [`Send`] and [`Sync`].
+///
+/// Using `SharedLazyData` as an extractor will not initialize the data; [`get`](Self::get) must be
+/// used.
+///
+/// [`LazyData`]: crate::extract::LazyData
+pub struct LazyDataShared<T> {
+    inner: Arc<LazyDataSharedInner<T>>,
 }
 
-struct LazyDataInner<T> {
+struct LazyDataSharedInner<T> {
     cell: OnceCell<T>,
-    fut: Cell<Option<LocalBoxFuture<'static, T>>>,
+    fut: Mutex<Option<BoxFuture<'static, T>>>,
 }
 
-impl<T> Clone for LazyData<T> {
+impl<T> Clone for LazyDataShared<T> {
     fn clone(&self) -> Self {
         Self {
             inner: self.inner.clone(),
@@ -30,31 +34,31 @@ impl<T> Clone for LazyData<T> {
     }
 }
 
-impl<T: fmt::Debug> fmt::Debug for LazyData<T> {
+impl<T: fmt::Debug> fmt::Debug for LazyDataShared<T> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         let Self { inner } = self;
-        let LazyDataInner { cell, fut: _ } = &**inner;
+        let LazyDataSharedInner { cell, fut: _ } = &**inner;
 
-        f.debug_struct("LazyData")
+        f.debug_struct("SharedLazyData")
             .field("cell", &cell)
-            .field("fut", &"<impl Future>")
+            .field("fut", &"..")
             .finish()
     }
 }
 
-impl<T> LazyData<T> {
+impl<T> LazyDataShared<T> {
     /// Constructs a new `LazyData` extractor with the given initialization function.
     ///
     /// Initialization functions must return a future that resolves to `T`.
-    pub fn new<F, Fut>(init: F) -> LazyData<T>
+    pub fn new<F, Fut>(init: F) -> LazyDataShared<T>
     where
         F: FnOnce() -> Fut,
-        Fut: Future<Output = T> + 'static,
+        Fut: Future<Output = T> + Send + 'static,
     {
         Self {
-            inner: Rc::new(LazyDataInner {
+            inner: Arc::new(LazyDataSharedInner {
                 cell: OnceCell::new(),
-                fut: Cell::new(Some(Box::pin(init()))),
+                fut: Mutex::new(Some(Box::pin(init()))),
             }),
         }
     }
@@ -64,7 +68,7 @@ impl<T> LazyData<T> {
         self.inner
             .cell
             .get_or_init(|| async move {
-                match self.inner.fut.take() {
+                match &mut *self.inner.fut.lock().await {
                     Some(fut) => fut.await,
                     None => panic!("LazyData instance has previously been poisoned"),
                 }
@@ -73,19 +77,19 @@ impl<T> LazyData<T> {
     }
 }
 
-impl<T: 'static> FromRequest for LazyData<T> {
+impl<T: 'static> FromRequest for LazyDataShared<T> {
     type Error = Error;
     type Future = Ready<Result<Self, Error>>;
 
     #[inline]
     fn from_request(req: &HttpRequest, _: &mut dev::Payload) -> Self::Future {
-        if let Some(lazy) = req.app_data::<LazyData<T>>() {
+        if let Some(lazy) = req.app_data::<LazyDataShared<T>>() {
             ready(Ok(lazy.clone()))
         } else {
             debug!(
-                "Failed to extract `LazyData<{}>` for `{}` handler. For the Data extractor to work \
-                correctly, wrap the data with `LazyData::new()` and pass it to `App::app_data()`. \
-                Ensure that types align in both the set and retrieve calls.",
+                "Failed to extract `SharedLazyData<{}>` for `{}` handler. For the Data extractor to \
+                work correctly, wrap the data with `SharedLazyData::new()` and pass it to \
+                `App::app_data()`. Ensure that types align in both the set and retrieve calls.",
                 core::any::type_name::<T>(),
                 req.match_name().unwrap_or_else(|| req.path())
             );
@@ -111,12 +115,14 @@ mod tests {
 
     use super::*;
 
+    static_assertions::assert_impl_all!(LazyDataShared<()>: Send, Sync);
+
     #[actix_web::test]
     async fn lazy_data() {
         let app = init_service(
             App::new()
-                .app_data(LazyData::new(|| async { 10usize }))
-                .service(web::resource("/").to(|_: LazyData<usize>| HttpResponse::Ok())),
+                .app_data(LazyDataShared::new(|| async { 10usize }))
+                .service(web::resource("/").to(|_: LazyDataShared<usize>| HttpResponse::Ok())),
         )
         .await;
         let req = TestRequest::default().to_request();
@@ -125,44 +131,52 @@ mod tests {
 
         let app = init_service(
             App::new()
-                .app_data(LazyData::new(|| async {
+                .app_data(LazyDataShared::new(|| async {
                     actix_web::rt::time::sleep(Duration::from_millis(40)).await;
-                    10usize
+                    10_usize
                 }))
-                .service(web::resource("/").to(|_: LazyData<usize>| HttpResponse::Ok())),
+                .service(
+                    web::resource("/").to(|lazy_num: LazyDataShared<usize>| async move {
+                        if *lazy_num.get().await == 10 {
+                            HttpResponse::Ok()
+                        } else {
+                            HttpResponse::InternalServerError()
+                        }
+                    }),
+                ),
         )
         .await;
         let req = TestRequest::default().to_request();
         let resp = call_service(&app, req).await;
-        assert_eq!(resp.status(), StatusCode::OK);
+        assert_eq!(StatusCode::OK, resp.status());
 
         let app = init_service(
             App::new()
-                .app_data(LazyData::new(|| async { 10u32 }))
-                .service(web::resource("/").to(|_: LazyData<usize>| HttpResponse::Ok())),
+                .app_data(LazyDataShared::new(|| async { 10u32 }))
+                .service(web::resource("/").to(|_: LazyDataShared<usize>| HttpResponse::Ok())),
         )
         .await;
         let req = TestRequest::default().to_request();
         let resp = call_service(&app, req).await;
-        assert_eq!(resp.status(), StatusCode::INTERNAL_SERVER_ERROR);
+        assert_eq!(StatusCode::INTERNAL_SERVER_ERROR, resp.status());
     }
 
     #[actix_web::test]
     async fn lazy_data_web_block() {
         let app = init_service(
             App::new()
-                .app_data(LazyData::new(|| async {
+                .app_data(LazyDataShared::new(|| async {
                     web::block(|| std::thread::sleep(Duration::from_millis(40)))
                         .await
                         .unwrap();
 
                     10usize
                 }))
-                .service(web::resource("/").to(|_: LazyData<usize>| HttpResponse::Ok())),
+                .service(web::resource("/").to(|_: LazyDataShared<usize>| HttpResponse::Ok())),
         )
         .await;
         let req = TestRequest::default().to_request();
         let resp = call_service(&app, req).await;
-        assert_eq!(resp.status(), StatusCode::OK);
+        assert_eq!(StatusCode::OK, resp.status());
     }
 }